+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.8.1] ??? 
+### Changed
+- The CLI ``pathfinder`` tool now wraps the search.py module's ``path_finder`` function. 
+
 ### Fixed
 - Github Actions now caching dependencies for faster builds.